+++ Packages/com.alelievr.mixture/Runtime/Nodes/Procedural/Cylinder.cs
@@ -1,4 +1,4 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 using GraphProcessor;
 namespace Mixture
 {
@@ -6,12 +6,12 @@ namespace Mixture
 Generates a line pattern. In 3D this node generate cylinders using a signed distance field function.
 ")]
 
-	[System.Serializable, NodeMenuItem("Procedural/Lines")]
-	public class Lines : FixedShaderNode
+	[System.Serializable, NodeMenuItem("Procedural/Cylinder")]
+	public class Cylinder : FixedShaderNode
 	{
-		public override string name => "Lines";
+		public override string name => "Cylinder";
 
-		public override string shaderName => "Hidden/Mixture/Lines";
+		public override string shaderName => "Hidden/Mixture/Cylinder";
 
 		public override bool displayMaterialInspector => true;