+++ CHANGELOG.md
@@ -8,11 +8,13 @@ All notable changes to this project template will be documented in this file.
 
 https://user-images.githubusercontent.com/6877923/123006036-64e2e780-d3b7-11eb-922e-018994b32da5.mov
 
-## 0.3.1
+- Added conversion methods for Vector3 and Vector2 to Vector4.
+- Added a popup to select which preview texture is visible in the output node.
 
 ### Fixed
 - Fixed missing nodes in the list of nodes proposed when dragging an edge.
-
+- Fixed gamma mismatch when exporting a 2D texture with the External Output node.
+- Fixed incorrect create node list when opened from an edge.
 
 ## 0.3.0