+++ CHANGELOG.md
@@ -22,6 +22,7 @@ https://user-images.githubusercontent.com/6877923/123006036-64e2e780-d3b7-11eb-9
 - Fixed incorrect create node list when opened from an edge.
 - Fixed processing of nodes not connected to the output in realtime mixtures.
 - Fixed a setting inheritance bug when a node inherited from a Texture node.
+- Fixed migration of old mixture graphs to the new settings system.
 
 ## 0.3.0